logitrack_backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
redis>=5.0,<7.0
strawberry-graphql[fastapi]>=0.220,<1.0
pydantic-settings>=2.2,<3.0
structlog>=24.1,<26.0
alembic>=1.13,<2.0
